(MODULES-5844) Move all code to dsc_lite namespace
This commit moves the PowerShellManager class and dependent files from
the `dsc` namespace to the `dsc_lite` namespace. Where it was possible,
template files were moved closer to the classes that used them. In all
cases a namespace was added to all types and classes to differentiate
the new module from any prior dsc module installs. Changes include:

 * Ruby PuppetX::Dsc -> PuppetX::DscLite
 * Ruby PuppetX::PuppetLabs::Dsc -> PuppetX::PuppetLabs::DscLite
 * renaming Puppet type base_dsc -> base_dcslite
 * using a dsc_lite path to namespace in the filesystem
 * updating existing wrappers (like the fake resource) to use the
   new names and paths