(MODULES-5842) Implement Generic DSC Invoker
This commit adds the `dsc` type and provider, which allows executing
DSC Resources without the build system used by the previous pupppet dsc
module.

This is accomplished by adding a new type that accepts a DSC Resource
Name and Module Name, and a un-typed hash of properties. This is passed
as is to `Invoke-DscResource`.

Unit tests for the type are added, manual verification of these changes
are to be done as automated tests are to be added in subsequent tickets