(MODULES-6323) EmbeddedInstance support
This commit adds support for EmbeddedInstances (complex types and
CimInstances) when using `Invoke-DscResource` with the generic `dsc`
type.

This commit also updates the `Usage` section of the README.md to reflect
the usage of the `dsc` type. It does not attempt to update the entire
README to reflect the changes compared to the dsc module, as that is out
of scope for this ticket.