(MODULES-5845) Add DSC Resource ModuleSpecification support
This commit changes `dsc_resource_module_name` to `dsc_resource_module`, and
updates it to accept both a string or a hash with two specific key/value
pairs, which loosely follows the `ModuleSpecification` format of the
`Invoke-DscResource` `ModuleName` parameter. A user can now specify
either the name of the DSC Resource Module *or* the name of the DSC
Resource Module and the version of the module to use. The name of the
DSC Resource module can either be the name of the module or the physical
path to the module.

This commit updates all doc and acceptance test references, adds unit
tests for the type validation, as well as adding an example for usage
in the readme.