(MODULES-6572) Refactor the acceptance tests
Prior to this commit the acceptance test suite files for the fake
dsc resource were split into two subfolders which did not match the
way that powershell modules are actually installed on disk. This
commit updates the layout to more accurately represent how powershell
modules are installed so that, when copied over for the tests, the
files represent what we might actually find on a target system.

A result of this movement was that several tests broke for two
primary reasons:

1. Several tests were referencing the generated type for the dsc
  resource instead of using dsc_lite to call that resource. This
  was not apparent until the pathing was updated and thus broke
  this behavior.
2. Many of the tests referenced a path to a 'dsc' module on the
  target system even though this project should install to 'dsc_lite'.

This commit ensures that all tests pass with the updated paths by
resolving these issues.

Additionally, it was discovered that for _some_ reason, placing the
dsc_lite resources in manifests *after* the declaration of the reboot
resource caused all of the reboot tests to fail. Simply moving them
to _before_ the declaration of the reboot tests caused them to pass.

@jpogran mentioned that this might be related to the initial
implementation of the reboot notification code. Resolving it was out
of scope for this PR but future discussions should determine whether
a ticket is needed.