(MODULES-6572) Add windows_agents helper
Prior to this commit every tests file needed to add confine
statements to prevent the tests running against non-windows
machines because those runs would always fail.

This commit adds the `windows_agents` method to the dsc
utils file allowing the tests to reference windows_agents
and skip the confine statements.

In the future, when this test suite is converted to beaker
rspec, this method will live in the acceptance_helper file.