(MAINT) Set reboot tests to pending
It came to our attention while implementing MODULES-6572 that the reboot tests
might not actually be functional. The ticket for implementing the functionality
being tested here is laid out in MODULES-6569.

This ticket sets these tests to pending as the functionality is not currently
expected to work and the implementation is slated for post 1.0.0.