(MODULES-7179) Change created to invoked
This commit changes the message that Puppet returns when dsc_lite
resources are run from "created" to "invoked". This is because at the
moment, when we invoke a dsc_lite resource, Puppet does not know if the
resource we want to manage has actually been created, or if in reality
it has simply been modified, or even deleted. Changing to invoked is
more ambiguous, but accurately reflects the idea that we don't truly
know what operation dsc has accomplished for us.