(MODULES-8175) Add safety to new-pscredential helper
Prior to this commit the helper function `new-pscredential`
in the template used to invoke DSC resources attempted to
cast an incoming value for password to a string. When passed
a hashtable, it was unable to extract the password for use
and instead set the password as the typename for the object.

This happened when the catalog was compiled on a Puppet 5
master to apply to a Puppet 6 agent or when compiled on a
Puppet 6 master for a Puppet 5 agent and specifying the
password using the sensitive datatype.

This problem was introduced in MODULES-8597 when the tags
for rich data serialization were changed from `__pcore_type_`
and `__pcore_value__` to the shorter `__ptype` and `__pvalue`.
After this change, the master and agent both expect different
keys for the sensitive data type, which is why this problem
only shows when the master and agent are mismatched versions.

This commit modifies the `new-pscredential` helper function
to include handling for both strings and hashtables, and
will extract the password if possible and error loudly if
not.