(MODULES-8175) Munge Properties Hash for Sensitive
Prior to this commit the module did not handle sensitive values in the
property hash when the master and agent were mismatched major versions -
for example, master 5.x and agent 6.x or vice versa. This problem stems
from the change in the core puppet language for handling rich data types
and caused the agent to be unable to properly deserialize the objects
marked sensitive.

This commit adds a munge_sensitive_hash method and uses it in the munging
of the properties hash itself, ensuring that sensitive datatypes will
be appropriately marked during an agent run.