(MODULES-8856) Redact debug output
Prior to this commit the provider would output the full PowerShell
log to be executed during debug runs. This was to ensure that we
could verify exactly what would be executed. However, this also
meant that credentials and other secrets could leak into the debug
output, and thus also into PuppetDB logs.

This commit modifies the provider and a helper to inject a PowerShell
comment after a secret as `# PuppetSensitive`. This does not interfere
with passing the data to PowerShell, but does allow us to redact the
sensitive information using a regex matcher prior to emitting via
debug logging.

That redaction is handled by a new class method on the provider called
`redact_content`.

This commit also includes updated spec tests to verify that sensitive
data, when present, can be both unwrapped *and* redacted prior to being
emitted.