(MODULES-9086) Increase pipe timeout to 180
This commit increases the timeout for waiting on the named
pipe to open. This resolves isssues where servers under
heavy load do not open the pipe quickly enough.

NOTE: The code for the PowerShell Code Manager in this
module is out of date with fixes distributed to other
modules and should be updated in a separate ticket.