(MODULES-10471) Allow namevar to have special chars
Prior to this commit, the namevar in the dsc type allowed a subset of
characters. This was not a functional limit and did not need to be in
place. This commit removes this restriction and ensure the name is a
string. This commit also fixes the spec tests for the `resource_name` as
they were only testing the `name` parameter.