(CAT-1776) Implement reusable workflows
This PR aims to bring the DSC_lite tool workflow structure in line with
the rest of the DevX toolset. This change implements some of the modules
reusable workflows for consistency, as we officially recognise DSC_lite
as having a module-like structure and function. Here is a list of
changes made:

- Removed the `auto_release.yml` workflow as it is no longer needed
- Implement the reusable release_prep workflow to replace it
- Removed the `spec.yml` workflow as it is no longer needed
- Removed the `stale.yml` workflow as it is no longer needed
- Implement the reusable mend workflow to collect internal data
- Adjust the reusable labeller workflow to label PRs