(CAT-1869) - Add configurable dsc_timeout param for dsc type
Prior to this commit, we set a default timeout of 20 minutes for the dsc
type, and this was not configurable.

This change adds a new "dsc_timeout" param to the dsc type, which
accepts an integer value for the number of seconds execution should wait
before exiting.

Users can now configure this timeout through their manifests, same as
you would with any other property see below:
```
dsc { "Running Script":
  resource_name => 'Script',
  module        => 'PSDesiredStateConfiguration',
  dsc_timeout   => 10,
  properties    => {
    getscript  => 'Start-Sleep -Seconds 5',
    testscript => '$false',
    setscript  => 'Start-Sleep -Seconds 5',
  },
}
```