(Bug) - Fix timeout matcher
Ths commit fixes the timeout matcher to more accurately match a
legitimate dsc timeout. Before, if output was nil for any other reason
dsc_lite would assume timeout which is not strictly correct. There is a
specific error message returned on dsc timeout which we should strive to
match instead.

Additionally, moved the logic to only execute for nodes with windows
powershell support as we can only pass a timeout to powershell on these
nodes.